Separate summary component, strings
