+++ src/components/Player/Pages/Overview/Overview.jsx
@@ -20,144 +20,13 @@ import Container from 'components/Container';
 import playerMatchesColumns from 'components/Player/Pages/Matches/playerMatchesColumns';
 import { playerHeroesOverviewColumns } from 'components/Player/Pages/Heroes/playerHeroesColumns';
 import { playerPeersOverviewColumns } from 'components/Player/Pages/Peers/playerPeersColumns';
-import {
-  isRadiant,
-  sum,
-  formatSeconds,
-  abbreviateNumber,
-} from 'utility';
 import { defaultPlayerMatchesOptions } from 'actions/player/playerMatchesActions';
-import { Link } from 'react-router';
-import heroes from 'dotaconstants/build/heroes.json';
 import styles from './Overview.css';
+import SummOfRecMatches from './Summary';
 
-const MAX_MATCHES_ROWS = 20;
+export const MAX_MATCHES_ROWS = 20;
 const MAX_HEROES_ROWS = 10;
 const MAX_PEERS_ROWS = 5;
-
-const SummOfRecMatches = ({ matchesLoading, matchesError, matchesData }) => {
-  // initial values
-  const data = {
-    kills: [],
-    deaths: [],
-    assists: [],
-    xp_per_min: [],
-    gold_per_min: [],
-    hero_damage: [],
-    hero_healing: [],
-    last_hits: [],
-    duration: [],
-
-    wins: [],
-  };
-  const computed = {};
-
-  let winrate = 0;
-  let kla = 0;
-
-  if (!matchesLoading && !matchesError) {
-    const dataKeys = Object.keys(data);
-
-    for (let i = 0; i < MAX_MATCHES_ROWS; i += 1) {
-      dataKeys.map((key) => {
-        if (key === 'wins') {
-          data.wins.push(matchesData[i].radiant_win === isRadiant(matchesData[i].player_slot));
-        } else {
-          data[key].push(matchesData[i][key]);
-        }
-
-        return null;
-      });
-    }
-
-    dataKeys.map((key) => {
-      if (key !== 'wins') {
-        const avg = data[key].reduce(sum, 0) / MAX_MATCHES_ROWS;
-        const max = Math.max(...data[key]);
-        const maxMatch = matchesData.find(match => match[key] === max);
-
-        let color;
-
-        switch (key) {
-          case 'kills':
-            color = 'green';
-            break;
-          case 'deaths':
-            color = 'red';
-            break;
-          case 'assists':
-            color = 'lightGray';
-            break;
-          case 'gold_per_min':
-            color = 'golden';
-            break;
-          default:
-            color = false;
-        }
-
-        computed[key] = {
-          avg,
-          color,
-          max: {
-            value: max,
-            matchId: maxMatch.match_id,
-            heroId: maxMatch.hero_id,
-            // ended: maxMatch.start_time + maxMatch.duration,
-          },
-        };
-      }
-
-      return null;
-    });
-
-    winrate = Number((data.wins
-    .filter(Boolean)
-    .reduce(sum, 0) * 100 / MAX_MATCHES_ROWS)
-      .toFixed(2));
-    kla = Number(((computed.kills.avg + computed.assists.avg) / (computed.deaths.avg + 1)).toFixed(2));
-  }
-
-  return (
-    <Container
-      title="Summary of Recent Matches"
-      className={styles.summaryContainer}
-      loading={matchesLoading}
-      error={matchesError}
-    >
-      <div>
-        <ul>
-          <li>
-            <span>winrate</span>
-            <p>{winrate}%</p>
-          </li>
-          <li>
-            <span>kla</span>
-            <p>{kla}</p>
-          </li>
-          {Object.keys(computed).map((key) => {
-            const c = computed[key];
-
-            return (
-              <li key={key}>
-                <span>{key}</span>
-                <p style={{ color: styles[c.color] }}>
-                  {key === 'duration' ? formatSeconds(c.avg) : abbreviateNumber(c.avg)}
-                  &nbsp;
-                  <span>{key === 'duration' ? formatSeconds(c.max.value) : abbreviateNumber(c.max.value)}
-                    &nbsp;
-                    <Link to={`matches/${c.max.matchId}`}>
-                      <img src={`${API_HOST}${heroes[c.max.heroId].icon}`} alt={heroes[c.max.heroId].localized_name} />
-                    </Link>
-                  </span>
-                </p>
-              </li>
-            );
-          })}
-        </ul>
-      </div>
-    </Container>
-  );
-};
 
 const Overview = ({
   matchesData,
