+++ src/lang/en-US.json
@@ -292,6 +292,8 @@
   "heading_region": "Region",
   "heading_patch": "Patch",
   "heading_win_rate": "Win Rate",
+  "heading_avg_and_max": "Average and maximum values",
+  "subheading_avg_and_max": "of the recent %d matches",
   
   "hero_pro_tab": "Professional",
   "hero_public_tab": "Public",
